% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{dist_repro}
\alias{dist_repro}
\title{dist_repro}
\usage{
dist_repro(eigenvalue, Theta, Cx, Fx, K)
}
\arguments{
\item{eigenvalue}{Output of DEEF function, eigenvalue}

\item{Theta}{Output of DEEF function, Theta}

\item{Cx}{Output of DEEF function, Cx}

\item{Fx}{Output of DEEF function, Fx}

\item{K}{the number of the used top theta coordinates}
}
\value{
samply by grid  reconstructed probability matrix
}
\description{
This Function is to reconstruct distribution set from the output of DEEF
}
\examples{
# See the tutorial document at "https://github.com/DaigoOkada/deef"
}
\keyword{NA}
