% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{Distset2D}
\alias{Distset2D}
\title{Distset2D}
\format{list contains the four objects
\describe{
\item{mu_sd}{900 by 2 matrix.The mean and sd of 900 member normal distributions.}
\item{ip_mat}{900 by 900 functional inner matrix.}
\item{grid}{The 10,000 grid values in discretization}
\item{P}{900 by 10,000 probability matrix.}
}}
\usage{
Distset2D
}
\description{
This dataset of distribution set 2D
}
\keyword{datasets}
