% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DEEF}
\alias{DEEF}
\title{DEEF}
\usage{
DEEF(disP, ip_mat = NULL)
}
\arguments{
\item{disP}{sample by grid plobability matrix, rowsums should be all 1.}

\item{ip_mat}{optional. if sample by sample matrix are given, it is used as innerproduct matrix.}
}
\value{
list that contains eigenvalue,Theta,Cx,Fx
}
\description{
This Function is to conduct DEEF method
}
\examples{

}
\keyword{NA}
